Arista,Cisco: fix handling of local-preference (#6089)
Local preference is a u32, so cannot be represented as an int. We kind of
handled this, but not consistently.

Also delete IOS-XR remnants from both vendors.

Fix batfish/batfish#6077.

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/routing_policy/expr/DecrementLocalPreference.java
@@ -11,7 +11,7 @@ public final class DecrementLocalPreference extends LongExpr {
   @JsonCreator
   private DecrementLocalPreference() {}
 
-  public DecrementLocalPreference(int subtrahend) {
+  public DecrementLocalPreference(long subtrahend) {
     _subtrahend = subtrahend;
   }
